Replace `long long` with plain `long`
We don't really need the extra digits from long long, it was
just a misdiagnosis made during the stream based on prints being
done one "integers".

+++ 06/lanternfish.h
@@ -1,3 +1,3 @@
-long long* parse_countdown(int* initial_states, int len_states);
-void advance_day(long long* countdowns);
-long long count_after_n_days(long long* countdowns, int days);
+long* parse_countdown(int* initial_states, int len_states);
+void advance_day(long* countdowns);
+long count_after_n_days(long* countdowns, int days);
